server/src/test/java/toxiproxy/api/controller/ToxiproxyAPIControllerExceptionHandlerTest.java
package toxiproxy.api.controller;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertNull;

@ExtendWith(MockitoExtension.class)
public class ToxiproxyAPIControllerExceptionHandlerTest {

  @InjectMocks private ToxiproxyAPIControllerExceptionHandler exceptionHandler;

  @Test
  void handleNotFoundException() {
    ResponseEntity responseEntity = exceptionHandler.handleNotFoundException(null);

    assertNotNull(responseEntity);
    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
    assertNull(responseEntity.getBody());
  }
}
